fix:quilleditor をコンポーネントとして切り出す
